refactor: auth API 레이어 정리
