+++ src/entities/auth/api/logout-auth.api.ts
@@ -1,4 +1,4 @@
-import { request } from "../core";
+import { request } from "@/shared/api/core";
 
 export const logoutAuthAPI = (): Promise<void> =>
   request<void>({
