+++ src/entities/index.ts
@@ -1,3 +1,4 @@
+export * from "./auth";
 export * from "./connection";
 export * from "./node";
 export * from "./workflow";
