+++ src/shared/api/index.ts
@@ -1,5 +1,4 @@
 export * from "./client";
-export * from "./auth";
 export * from "./execution";
 export * from "./oauth-token";
 export * from "./template";
